corrected entities creation position
