--- generators/entities/generator.spec.mjs
@@ -1,29 +1,0 @@
-import { beforeAll, describe, expect, it } from 'vitest';
-
-import { helpers, lookups } from '#test-utils';
-
-const SUB_GENERATOR = 'entities';
-const BLUEPRINT_NAMESPACE = `jhipster:${SUB_GENERATOR}`;
-
-describe('SubGenerator entities of native JHipster blueprint', () => {
-  describe('run', () => {
-    let result;
-    beforeAll(async function () {
-      result = await helpers
-        .create(BLUEPRINT_NAMESPACE)
-        .withOptions({
-          reproducible: true,
-          defaults: true,
-          baseName: 'jhipster',
-          ignoreNeedlesError: true,
-          blueprint: 'native',
-        })
-        .withLookups(lookups)
-        .run();
-    });
-
-    it('should succeed', () => {
-      expect(result.getStateSnapshot()).toMatchSnapshot();
-    });
-  });
-});
