--- generators/entities/index.mjs
@@ -1,1 +1,0 @@
-export { default } from './generator.mjs';
