+++ generators/server/generator.mjs
@@ -151,6 +151,67 @@ class `,
 
   get [BaseApplicationGenerator.POST_WRITING_ENTITIES]() {
     return this.asPostWritingTaskGroup({
+      async entities({ application: { reactive, databaseTypeSql }, entities }) {
+        for (const { name } of entities.filter(({ builtIn, embedded }) => !builtIn && !embedded)) {
+          // Use entity from old location for more complete data.
+          const entity = this.sharedData.getEntity(name);
+          if (!entity) {
+            this.log.warn(`Skipping entity generation, use '--with-entities' flag`);
+            // eslint-disable-next-line no-continue
+            continue;
+          }
+          this.editFile(`${JAVA_MAIN_SOURCES_DIR}/${entity.entityAbsoluteFolder}/web/rest/${entity.entityClass}Resource.java`, content =>
+            content
+              .replaceAll(
+                `@PathVariable(value = "${entity.primaryKey.name}", required = false) final ${entity.primaryKey.type} ${entity.primaryKey.name}`,
+                `@PathVariable(name = "${entity.primaryKey.name}", value = "${entity.primaryKey.name}", required = false) final ${entity.primaryKey.type} ${entity.primaryKey.name}`,
+              )
+              .replaceAll(
+                `@PathVariable ${entity.primaryKey.type} ${entity.primaryKey.name}`,
+                `@PathVariable("${entity.primaryKey.name}") ${entity.primaryKey.type} ${entity.primaryKey.name}`,
+              )
+              .replaceAll(
+                `@RequestParam(required = false, defaultValue = "false") boolean eagerload`,
+                `@RequestParam(name = "eagerload", required = false, defaultValue = "false") boolean eagerload`,
+              )
+              .replaceAll(
+                `@RequestParam(required = false, defaultValue = "true") boolean eagerload`,
+                `@RequestParam(name = "eagerload", required = false, defaultValue = "true") boolean eagerload`,
+              ),
+          );
+
+          if (!reactive && databaseTypeSql && entity.containsBagRelationships) {
+            this.editFile(
+              `${JAVA_MAIN_SOURCES_DIR}${entity.entityAbsoluteFolder}/repository/${entity.entityClass}RepositoryWithBagRelationshipsImpl.java`,
+              contents =>
+                contents.replace(
+                  'import org.springframework.beans.factory.annotation.Autowired;',
+                  'import javax.persistence.PersistenceContext;',
+                ),
+              contents => contents.replace('@Autowired', '@PersistenceContext'),
+            );
+          }
+
+          if (reactive && databaseTypeSql) {
+            this.editFile(
+              `${JAVA_MAIN_SOURCES_DIR}${entity.entityAbsoluteFolder}/repository/${entity.entityClass}RepositoryInternalImpl.java`,
+              contents =>
+                contents.replace(
+                  'import reactor.core.publisher.Flux;',
+                  `import org.springframework.stereotype.Component;
+import reactor.core.publisher.Flux;`,
+                ),
+              contents =>
+                contents.replace(
+                  '\nclass ',
+                  `
+@Component
+class `,
+                ),
+            );
+          }
+        }
+      },
       async jsonFilter({ entities }) {
         // include user entity.
         const targetEntities = [...entities.filter(({ builtIn, embedded }) => !builtIn && !embedded), this.sharedData.getEntity('User')];
